Support filtering on a closed bound for the statistics.
By default, this is an open bound as we usually want to find genes that have
statistics above a threshold (e.g., above zero for the residual variances or
log-fold changes). However, it is possible that we might want a closed bound to
retain genes that are exactly on the threshold; this option enables that.